remove adjust height, fix bug when excelPanel in ViewPager
