+++ app/build.gradle
@@ -26,6 +26,6 @@ dependencies {
     })
     compile project(':excelpanel')
     compile 'com.android.support:appcompat-v7:24.0.0'
-    compile 'com.kelin.scrollablepanel:library:1.2.0'
+    compile 'com.android.support:design:24.0.0'
     testCompile 'junit:junit:4.12'
 }
