+++ app/src/main/res/layout/activity_main.xml
@@ -1,23 +1,27 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/activity_main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:orientation="vertical"
     tools:context="cn.zhouchaoyuan.excelpaneldemo.MainActivity">
 
-    <cn.zhouchaoyuan.excelpanel.ExcelPanel
-        android:id="@+id/content_container"
+    <android.support.design.widget.TabLayout
+        android:id="@+id/tab_layout"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:left_cell_width="@dimen/room_status_cell_length"
-        app:normal_cell_width="@dimen/room_status_cell_length"
-        app:top_cell_height="@dimen/room_status_cell_length" />
+        android:layout_height="50dp"
+        android:background="@color/bg_color"
+        app:tabIndicatorColor="@color/tab_indicator_color"
+        app:tabMode="scrollable"
+        app:tabTextColor="@color/white">
 
-    <ProgressBar
-        android:id="@+id/progress"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center" />
-</FrameLayout>
+    </android.support.design.widget.TabLayout>
+
+    <android.support.v4.view.ViewPager
+        android:id="@+id/pager"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1"></android.support.v4.view.ViewPager>
+</LinearLayout>
