+++ excelpanel/src/main/java/cn/zhouchaoyuan/excelpanel/LeftRecyclerViewAdapter.java
@@ -2,7 +2,6 @@ package cn.zhouchaoyuan.excelpanel;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
-import android.util.Pair;
 import android.view.ViewGroup;
 
 import java.util.List;
@@ -41,9 +40,6 @@ public class LeftRecyclerViewAdapter<L> extends RecyclerViewAdapter<L> {
     public void onBindNormalViewHolder(RecyclerView.ViewHolder holder, int position) {
         if (excelPanelListener != null) {
             excelPanelListener.onBindLeftViewHolder(holder, position);
-            //use to adjust height
-            holder.itemView.setTag(ExcelPanel.TAG_KEY, new Pair<>(position, 0));
-            excelPanelListener.onAfterBind(holder, position, true, true);
         }
     }
 }
