--- excelpanel/src/main/java/cn/zhouchaoyuan/excelpanel/OnAddVerticalScrollListener.java
@@ -1,11 +1,0 @@
-package cn.zhouchaoyuan.excelpanel;
-
-import android.support.v7.widget.RecyclerView;
-
-/**
- * Created by zhouchaoyuan on 2016/12/12.
- */
-
-public interface OnAddVerticalScrollListener {
-    void addRecyclerView(RecyclerView recyclerView);
-}
